LPS-171250 Move ActionContainer and ThenContainer components to a new directory
