+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/ObjectAction/tabs/ActionContainer/ThenContainer.tsx
@@ -29,11 +29,11 @@ import {
 	ObjectsOptionsList,
 	fetchObjectDefinitionFields,
 	fetchObjectDefinitions,
-} from '../fetchUtil';
+} from '../../fetchUtil';
 
 import './ActionBuilder.scss';
 
-import {ActionError} from '..';
+import {ActionError} from '../..';
 
 interface ThenContainerProps {
 	errors: ActionError;
